only support dynamic library in v3.2.1
