+++ PLPlayerKit.podspec
@@ -9,7 +9,7 @@
 
 Pod::Spec.new do |s|
   s.name             = "PLPlayerKit"
-  s.version          = "3.2.0"
+  s.version          = "3.2.1"
   s.summary          = "Pili iOS video player SDK, RTMP, HLS video streaming supported."
   s.homepage         = "https://github.com/pili-engineering/PLPlayerKit"
   s.license          = 'Apache License, Version 2.0'
@@ -19,9 +19,9 @@ Pod::Spec.new do |s|
   s.platform     = :ios, '8.0'
   s.requires_arc = true
 
-  s.public_header_files = "Pod/Library/include/PLPlayerKit/*.h"
-  s.source_files = 'Pod/Library/include/**/*.[h|m]'
-  s.vendored_libraries   = 'Pod/Library/lib/*.a'
+  s.subspec "iphoneos" do |ss1|
+    ss1.vendored_framework = "Pod/Library/PLPlayerKit.framework"
+  end
 
   s.dependency 'HappyDNS', '~> 0.3.10'
   s.frameworks = ["UIKit", "Foundation", "MediaPlayer", "CoreAudio", "AudioToolbox", "Accelerate", "QuartzCore", "OpenGLES", "AVFoundation","CoreVideo","AVKit","CoreMedia","VideoToolbox","CoreTelephony"]
