+++ Pod/Library-Universal/PLPlayerKit.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module PLPlayerKit {
+  umbrella header "PLPlayerKit.h"
+
+  export *
+  module * { export * }
+}
