--- Pod/Library/include/PLPlayerKit/PLPlayerKit.h
@@ -1,17 +1,0 @@
-//
-//  PLPlayerKit.h
-//  PLPlayerKit
-//
-//  Created by 何昊宇 on 2017/5/15.
-//  Copyright © 2017年 Aaron. All rights reserved.
-//
-
-#ifndef PLPlayerKit_h
-#define PLPlayerKit_h
-
-#import "PLPlayer.h"
-#import "PLPlayerOption.h"
-#import "PLPlayerError.h"
-
-
-#endif /* PLPlayerKit_h */
